I: Watch preliminary work
Stopping for now watch, since it's unclear whether it supports ble peripheral
If it does, we should implement a standalone app for watch.
in this case we'd not support at all bridging with the iPhone as this would be only for older watch versions and it doesn't seem worth it.