I: Fix during meeting, one of the devices doesn't see peer Also, added tolerance to timers, which is said to be better for performance. We don't need high accuracy in any of the timers.
Re: meeting, it turns that one of the devices wasn't being able to read the validation characteristic
The reason: this device for some was reason was observing the peripheral of the other (even if we hadn't started advertising yet, app reinstall, etc) without any services, and then the other actually started advertising (with the service), the former device still saw no services. We had to implement the central's didModifyServices delegate method to trigger a discovery again, to see the service.